Organize the directory structure
