+++ src/ui/react/types.ts
@@ -1,0 +1,3 @@
+import React from "react";
+
+export type ReactType = typeof React;
